requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
python-dotenv
httpx
pytest
